local-llm-serve-guard/pyproject.toml
[project]
name = "local-llm-serve-guard"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "aiosqlite>=0.22.1",
    "fastapi>=0.135.1",
    "httpx>=0.28.1",
    "pyyaml>=6.0.3",
    "uvicorn>=0.41.0",
]
